Platforms/AMD/OverdriveBoard: add graphical console support
Add the modules that are needed to run the console in graphical
mode, i.e., when a GFX expansion card with a compatible option
ROM is inserted into the system.

Note that this includes the BGRT driver, which creates the BGRT ACPI
table on the fly, so update the hardcoded OEM ids as well.

Since the BGRT driver generates ACPI tables on the fly, set the OEM
vendor and table IDs to the same value we use in the static tables.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>