HiKeyDxe: Enable wifi
The wifi on the hikey is attached to mmc2 but the clock is
gated by the hi6553 PMIC which is memory mapped via the SSI
interface. Out of the box this means that when the kernel
attempts to communicate on dwmmc2 the bus hangs.

To avoid that hang we assure that the mmc2 isn't stuck in reset
and that the PMIC is generating a 32Khz clock and pumping it out
CLK32B by default. Once that is done, we renable the mmc2 in the
FDT and now the lan driver can bind.

Signed-off-by: Jeremy Linton <lintonrjeremy@gmail.com>