Platforms/Marvell: switch to generic ResetSystemRuntimeDxe
Switch to the generic implementation of the ResetSystem() runtime
service call, which is preferred over the one in EmbeddedPkg. This
involves selecting another DXE runtime driver, and given that the
Marvell platforms support PSCI these days, we can drop the existing
MvResetSystemLib entirely and switch to the shared PSCI based one.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>