drivers:BlockVariableDxe: Enable persistent variables in linux
Bypass the block layer and write the variables directly to the emmc
on the hikey by setting up a block read/write method that twiddles
the emmc controller directly.

This is a fairly large hack, as it defeats the purpose of the
BlockVariableDxe which is to translate between a block device and
the UEFI variable store. An idea that initially sounds good until
you consider the edk2 layering affects basically forces a huge
portion of the DXE layer to be runtime safe.

OTOH, it works...